[UPSTREAM] libarchive: check for a TEST_SLOW_HOST environment variable
When set we run fewer iterations or use smaller buffer sizes for the tests.
This is especially noticeable on emulated platforms: the
test_write_format_7zip_large_copy test takes ~106 seconds on a
pure-capability CHERI RISC-V QEMU by default, but with the TEST_SLOW_HOST
environment variable set we use a 1MB data size instead of 16MB and the
duration is now only ~7 seconds.

Additionally, the fuzz tests often time out when running them as part of
our CI setup on CHERI QEMU RISC-V. To avoid this problem we run 10
iterations of the test instead of 1000 when TEST_SLOW_HOST=1 is set.